fix: rename to base-cloud-formation ml-31
