+++ apps/backend/src/libs/modules/cloud-formation/libs/type/cloud-formation.type.ts
@@ -1,0 +1,8 @@
+import { type CreateStack } from "./types.js";
+
+type CloudFormation = {
+	create(payload: CreateStack): Promise<string>;
+	delete(meetingId: number): Promise<void>;
+};
+
+export { type CloudFormation };
