--- apps/backend/src/libs/modules/cloud-formation/libs/type/create-instance.types.ts
@@ -1,6 +1,0 @@
-type CreateInstance = {
-	id: number;
-	template: string;
-};
-
-export { type CreateInstance };
