+++ apps/backend/src/libs/modules/cloud-formation/libs/type/create-stack.types.ts
@@ -1,0 +1,6 @@
+type CreateStack = {
+	id: number;
+	template: string;
+};
+
+export { type CreateStack };
