+++ apps/backend/src/libs/modules/cloud-formation/libs/type/types.ts
@@ -1,1 +1,2 @@
-export { type CreateInstance } from "./create-instance.types.js";
+export { type CloudFormation } from "./cloud-formation.type.js";
+export { type CreateStack } from "./create-stack.types.js";
