+++ apps/backend/src/modules/meetings/meeting.service.ts
@@ -1,4 +1,4 @@
-import { type CloudFormation } from "~/libs/modules/cloud-formation/cloud-formation.module.js";
+import { type CloudFormation } from "~/libs/modules/cloud-formation/cloud-formation.js";
 import template from "~/libs/modules/cloud-formation/libs/templates/ec2-instance-template.json" with { type: "json" };
 import { HTTPCode } from "~/libs/modules/http/http.js";
 import { type Service } from "~/libs/types/types.js";
